requirements.txt
jieba==0.42.1
matplotlib==3.1.2
numpy==1.17.4
pillow==8.0.0
pandas==0.25.3
wordcloud==1.8.0
